Improve ip_address_check decorator
